+++ main/read.c
@@ -80,9 +80,9 @@ typedef struct sInputLineFposMap {
 
 typedef struct sNestedInputStreamInfo {
 	unsigned long startLine;
-	int startCharOffset;
+	long startCharOffset;
 	unsigned long endLine;
-	int endCharOffset;
+	long endCharOffset;
 } nestedInputStreamInfo;
 
 typedef struct sInputFile {
@@ -1028,8 +1028,8 @@ out:
 }
 
 extern void   pushNarrowedInputStream (const langType language,
-				       unsigned long startLine, int startCharOffset,
-				       unsigned long endLine, int endCharOffset,
+				       unsigned long startLine, long startCharOffset,
+				       unsigned long endLine, long endCharOffset,
 				       unsigned long sourceLineOffset)
 {
 	long p, q;
